+++ common/src/main/java/net/spell_engine/rpg_series/datagen/WeaponSkills.java
@@ -199,7 +199,7 @@ public class WeaponSkills {
         disrupt.sound = Sound.of(SpellEngineSounds.WEAPON_MACE_SMASH_IMPACT.id());
         spell.impacts = List.of(damage, disrupt);
 
-        SpellBuilder.Cost.cooldown(spell, 12);
+        SpellBuilder.Cost.cooldown(spell, 10);
         spell.cost.cooldown.attempt_duration = 1F;
 
         return new Entry(id, spell, title, description, null);
@@ -329,7 +329,7 @@ public class WeaponSkills {
         damage.sound = Sound.of(SpellEngineSounds.WEAPON_SPEAR_STAB.id());
         spell.impacts = List.of(damage);
 
-        SpellBuilder.Cost.cooldown(spell, 12);
+        SpellBuilder.Cost.cooldown(spell, 10);
 
         return new Entry(id, spell, title, description, null);
     }